+++ .storybook/webpack.config.js
@@ -1,0 +1,4 @@
+
+module.exports = function(...args) {
+  return require('/Users/caixiaobing/select/node_modules/_father@2.29.2@father/lib/doc/storybook-webpack')(...args);
+};