+++ es/Option.js
@@ -1,23 +1,7 @@
-import _classCallCheck from 'babel-runtime/helpers/classCallCheck';
-import _possibleConstructorReturn from 'babel-runtime/helpers/possibleConstructorReturn';
-import _inherits from 'babel-runtime/helpers/inherits';
-import React from 'react';
-import PropTypes from 'prop-types';
-
-var Option = function (_React$Component) {
-  _inherits(Option, _React$Component);
-
-  function Option() {
-    _classCallCheck(this, Option);
-
-    return _possibleConstructorReturn(this, (Option.__proto__ || Object.getPrototypeOf(Option)).apply(this, arguments));
-  }
-
-  return Option;
-}(React.Component);
-
-Option.propTypes = {
-  value: PropTypes.string
+/** This is a placeholder, not real render in dom */
+var Option = function Option() {
+  return null;
 };
+
 Option.isSelectOption = true;
 export default Option;