+++ es/SelectTrigger.js
@@ -1,196 +1,129 @@
-import _defineProperty from 'babel-runtime/helpers/defineProperty';
-import _objectWithoutProperties from 'babel-runtime/helpers/objectWithoutProperties';
-import _extends from 'babel-runtime/helpers/extends';
-import _classCallCheck from 'babel-runtime/helpers/classCallCheck';
-import _createClass from 'babel-runtime/helpers/createClass';
-import _possibleConstructorReturn from 'babel-runtime/helpers/possibleConstructorReturn';
-import _inherits from 'babel-runtime/helpers/inherits';
+function ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }
+
+function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(Object(source), true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(Object(source)).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }
+
+function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
+
+function _objectWithoutProperties(source, excluded) { if (source == null) return {}; var target = _objectWithoutPropertiesLoose(source, excluded); var key, i; if (Object.getOwnPropertySymbols) { var sourceSymbolKeys = Object.getOwnPropertySymbols(source); for (i = 0; i < sourceSymbolKeys.length; i++) { key = sourceSymbolKeys[i]; if (excluded.indexOf(key) >= 0) continue; if (!Object.prototype.propertyIsEnumerable.call(source, key)) continue; target[key] = source[key]; } } return target; }
+
+function _objectWithoutPropertiesLoose(source, excluded) { if (source == null) return {}; var target = {}; var sourceKeys = Object.keys(source); var key, i; for (i = 0; i < sourceKeys.length; i++) { key = sourceKeys[i]; if (excluded.indexOf(key) >= 0) continue; target[key] = source[key]; } return target; }
+
+import * as React from 'react';
 import Trigger from 'rc-trigger';
-import React from 'react';
-import PropTypes from 'prop-types';
-import classnames from 'classnames';
-import DropdownMenu from './DropdownMenu';
-import ReactDOM from 'react-dom';
-import { isSingleMode } from './util';
+import classNames from 'classnames';
 
-Trigger.displayName = 'Trigger';
-
-var BUILT_IN_PLACEMENTS = {
-  bottomLeft: {
-    points: ['tl', 'bl'],
-    offset: [0, 4],
-    overflow: {
-      adjustX: 0,
-      adjustY: 1
+var getBuiltInPlacements = function getBuiltInPlacements(dropdownMatchSelectWidth) {
+  // Enable horizontal overflow auto-adjustment when a custom dropdown width is provided
+  var adjustX = typeof dropdownMatchSelectWidth !== 'number' ? 0 : 1;
+  return {
+    bottomLeft: {
+      points: ['tl', 'bl'],
+      offset: [0, 4],
+      overflow: {
+        adjustX: adjustX,
+        adjustY: 1
+      }
+    },
+    bottomRight: {
+      points: ['tr', 'br'],
+      offset: [0, 4],
+      overflow: {
+        adjustX: adjustX,
+        adjustY: 1
+      }
+    },
+    topLeft: {
+      points: ['bl', 'tl'],
+      offset: [0, -4],
+      overflow: {
+        adjustX: adjustX,
+        adjustY: 1
+      }
+    },
+    topRight: {
+      points: ['br', 'tr'],
+      offset: [0, -4],
+      overflow: {
+        adjustX: adjustX,
+        adjustY: 1
+      }
     }
-  },
-  topLeft: {
-    points: ['bl', 'tl'],
-    offset: [0, -4],
-    overflow: {
-      adjustX: 0,
-      adjustY: 1
-    }
-  }
+  };
 };
 
-var SelectTrigger = function (_React$Component) {
-  _inherits(SelectTrigger, _React$Component);
+var SelectTrigger = function SelectTrigger(props, ref) {
+  var prefixCls = props.prefixCls,
+      disabled = props.disabled,
+      visible = props.visible,
+      children = props.children,
+      popupElement = props.popupElement,
+      containerWidth = props.containerWidth,
+      animation = props.animation,
+      transitionName = props.transitionName,
+      dropdownStyle = props.dropdownStyle,
+      dropdownClassName = props.dropdownClassName,
+      _props$direction = props.direction,
+      direction = _props$direction === void 0 ? 'ltr' : _props$direction,
+      _props$dropdownMatchS = props.dropdownMatchSelectWidth,
+      dropdownMatchSelectWidth = _props$dropdownMatchS === void 0 ? true : _props$dropdownMatchS,
+      dropdownRender = props.dropdownRender,
+      dropdownAlign = props.dropdownAlign,
+      getPopupContainer = props.getPopupContainer,
+      empty = props.empty,
+      getTriggerDOMNode = props.getTriggerDOMNode,
+      restProps = _objectWithoutProperties(props, ["prefixCls", "disabled", "visible", "children", "popupElement", "containerWidth", "animation", "transitionName", "dropdownStyle", "dropdownClassName", "direction", "dropdownMatchSelectWidth", "dropdownRender", "dropdownAlign", "getPopupContainer", "empty", "getTriggerDOMNode"]);
 
-  function SelectTrigger() {
-    var _ref;
+  var dropdownPrefixCls = "".concat(prefixCls, "-dropdown");
+  var popupNode = popupElement;
 
-    var _temp, _this, _ret;
-
-    _classCallCheck(this, SelectTrigger);
-
-    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
-      args[_key] = arguments[_key];
-    }
-
-    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_ref = SelectTrigger.__proto__ || Object.getPrototypeOf(SelectTrigger)).call.apply(_ref, [this].concat(args))), _this), _this.state = {
-      dropdownWidth: null
-    }, _this.setDropdownWidth = function () {
-      var visible = _this.props.visible;
-
-      if (visible) {
-        var width = ReactDOM.findDOMNode(_this).offsetWidth;
-        if (width !== _this.state.dropdownWidth) {
-          _this.setState({ dropdownWidth: width });
-        }
-      }
-    }, _this.getInnerMenu = function () {
-      return _this.popupMenu && _this.popupMenu.refs.menu;
-    }, _this.getPopupDOMNode = function () {
-      return _this.refs.trigger.getPopupDomNode();
-    }, _this.getDropdownElement = function (newProps) {
-      var props = _this.props;
-      return React.createElement(DropdownMenu, _extends({
-        ref: _this.saveMenu
-      }, newProps, {
-        prefixCls: _this.getDropdownPrefixCls(),
-        onMenuSelect: props.onMenuSelect,
-        onMenuDeselect: props.onMenuDeselect,
-        value: props.value,
-        firstActiveValue: props.firstActiveValue,
-        defaultActiveFirstOption: props.defaultActiveFirstOption,
-        dropdownMenuStyle: props.dropdownMenuStyle,
-        dropdownFooter: props.dropdownFooter,
-        dropdownHeader: props.dropdownHeader,
-        enableHideDropdownByClick: props.enableHideDropdownByClick,
-        onPopupVisibleChange: props.onDropdownVisibleChange
-      }));
-    }, _this.getDropdownTransitionName = function () {
-      var props = _this.props;
-      var transitionName = props.transitionName;
-      if (!transitionName && props.animation) {
-        transitionName = _this.getDropdownPrefixCls() + '-' + props.animation;
-      }
-      return transitionName;
-    }, _this.getDropdownPrefixCls = function () {
-      return _this.props.prefixCls + '-dropdown';
-    }, _this.saveMenu = function (menu) {
-      _this.popupMenu = menu;
-    }, _temp), _possibleConstructorReturn(_this, _ret);
+  if (dropdownRender) {
+    popupNode = dropdownRender(popupElement);
   }
 
-  _createClass(SelectTrigger, [{
-    key: 'componentDidMount',
-    value: function componentDidMount() {
-      this.setDropdownWidth();
-    }
-  }, {
-    key: 'componentDidUpdate',
-    value: function componentDidUpdate() {
-      this.setDropdownWidth();
-    }
-  }, {
-    key: 'render',
-    value: function render() {
-      var _popupClassName;
+  var builtInPlacements = React.useMemo(function () {
+    return getBuiltInPlacements(dropdownMatchSelectWidth);
+  }, [dropdownMatchSelectWidth]); // ===================== Motion ======================
 
-      var _props = this.props,
-          onPopupFocus = _props.onPopupFocus,
-          props = _objectWithoutProperties(_props, ['onPopupFocus']);
+  var mergedTransitionName = animation ? "".concat(dropdownPrefixCls, "-").concat(animation) : transitionName; // ======================= Ref =======================
 
-      var multiple = props.multiple,
-          visible = props.visible,
-          inputValue = props.inputValue,
-          dropdownAlign = props.dropdownAlign,
-          disabled = props.disabled,
-          showSearch = props.showSearch,
-          dropdownClassName = props.dropdownClassName,
-          dropdownStyle = props.dropdownStyle,
-          dropdownMatchSelectWidth = props.dropdownMatchSelectWidth;
-
-      var dropdownPrefixCls = this.getDropdownPrefixCls();
-      var popupClassName = (_popupClassName = {}, _defineProperty(_popupClassName, dropdownClassName, !!dropdownClassName), _defineProperty(_popupClassName, dropdownPrefixCls + '--' + (multiple ? 'multiple' : 'single'), 1), _popupClassName);
-      var popupElement = this.getDropdownElement({
-        menuItems: props.options,
-        onPopupFocus: onPopupFocus,
-        multiple: multiple,
-        inputValue: inputValue,
-        visible: visible
-      });
-      var hideAction = void 0;
-      if (disabled) {
-        hideAction = [];
-      } else if (isSingleMode(props) && !showSearch) {
-        hideAction = ['click'];
-      } else {
-        hideAction = ['blur'];
+  var popupRef = React.useRef(null);
+  React.useImperativeHandle(ref, function () {
+    return {
+      getPopupElement: function getPopupElement() {
+        return popupRef.current;
       }
-      var popupStyle = _extends({}, dropdownStyle);
-      var widthProp = dropdownMatchSelectWidth ? 'width' : 'minWidth';
-      if (this.state.dropdownWidth) {
-        popupStyle[widthProp] = this.state.dropdownWidth + 'px';
-      }
+    };
+  });
 
-      return React.createElement(
-        Trigger,
-        _extends({}, props, {
-          showAction: disabled ? [] : ['click'],
-          hideAction: hideAction,
-          ref: 'trigger',
-          popupPlacement: 'bottomLeft',
-          builtinPlacements: BUILT_IN_PLACEMENTS,
-          prefixCls: dropdownPrefixCls,
-          popupTransitionName: this.getDropdownTransitionName(),
-          onPopupVisibleChange: props.onDropdownVisibleChange,
-          popup: popupElement,
-          popupAlign: dropdownAlign,
-          popupVisible: visible,
-          getPopupContainer: props.getPopupContainer,
-          popupClassName: classnames(popupClassName),
-          popupStyle: popupStyle
-        }),
-        props.children
-      );
-    }
-  }]);
+  var popupStyle = _objectSpread({
+    minWidth: containerWidth
+  }, dropdownStyle);
 
-  return SelectTrigger;
-}(React.Component);
+  if (typeof dropdownMatchSelectWidth === 'number') {
+    popupStyle.width = dropdownMatchSelectWidth;
+  } else if (dropdownMatchSelectWidth) {
+    popupStyle.width = containerWidth;
+  }
 
-SelectTrigger.propTypes = {
-  onPopupFocus: PropTypes.func,
-  dropdownMatchSelectWidth: PropTypes.bool,
-  dropdownAlign: PropTypes.object,
-  visible: PropTypes.bool,
-  disabled: PropTypes.bool,
-  showSearch: PropTypes.bool,
-  dropdownClassName: PropTypes.string,
-  multiple: PropTypes.bool,
-  inputValue: PropTypes.string,
-  filterOption: PropTypes.any,
-  options: PropTypes.any,
-  prefixCls: PropTypes.string,
-  popupClassName: PropTypes.string,
-  children: PropTypes.any,
-  dropdownFooter: PropTypes.any,
-  dropdownHeader: PropTypes.any
+  return React.createElement(Trigger, Object.assign({}, restProps, {
+    showAction: [],
+    hideAction: [],
+    popupPlacement: direction === 'rtl' ? 'bottomRight' : 'bottomLeft',
+    builtinPlacements: builtInPlacements,
+    prefixCls: dropdownPrefixCls,
+    popupTransitionName: mergedTransitionName,
+    popup: React.createElement("div", {
+      ref: popupRef
+    }, popupNode),
+    popupAlign: dropdownAlign,
+    popupVisible: visible,
+    getPopupContainer: getPopupContainer,
+    popupClassName: classNames(dropdownClassName, _defineProperty({}, "".concat(dropdownPrefixCls, "-empty"), empty)),
+    popupStyle: popupStyle,
+    getTriggerDOMNode: getTriggerDOMNode
+  }), children);
 };
-export default SelectTrigger;
 
-
-SelectTrigger.displayName = 'SelectTrigger';
+var RefSelectTrigger = React.forwardRef(SelectTrigger);
+RefSelectTrigger.displayName = 'SelectTrigger';
+export default RefSelectTrigger;